requirements.txt
Flask
python-dotenv
langchain_google_genai
